backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
bcrypt==4.0.1
passlib[bcrypt]
python-dotenv==1.0.0
pydantic[email]==2.5.0
